// extensions/vscode/src/daemon/gen/ferretd/daemon/v1/daemon.pb.ts
export interface ApiVersion {
  major: number;
  minor: number;
}

export interface GetInfoRequest {
  clientApi?: ApiVersion | undefined;
}

export interface ServerInfo {
  version: string;
  instanceId: string;
  apiVersion?: ApiVersion | undefined;
}

export interface GetInfoResponse {
  serverInfo?: ServerInfo | undefined;
}

export interface ApiCompatibilityError {
  clientApi?: ApiVersion | undefined;
  serverApi?: ApiVersion | undefined;
}

export interface ShutdownRequest {
}

export interface ShutdownResponse {
}

function createBaseApiVersion(): ApiVersion {
  return { major: 0, minor: 0 };
}

export const ApiVersion: MessageFns<ApiVersion> = {
  encode(message: ApiVersion, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.major !== 0) {
      writer.uint32(8).uint32(message.major);
    }
    if (message.minor !== 0) {
      writer.uint32(16).uint32(message.minor);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApiVersion {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApiVersion();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.major = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.minor = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<ApiVersion>): ApiVersion {
    return ApiVersion.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ApiVersion>): ApiVersion {
    const message = createBaseApiVersion();
    message.major = object.major ?? 0;
    message.minor = object.minor ?? 0;
    return message;
  },
};
// ...
export const ApiCompatibilityError: MessageFns<ApiCompatibilityError> = {
  encode(message: ApiCompatibilityError, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.clientApi !== undefined) {
      ApiVersion.encode(message.clientApi, writer.uint32(10).fork()).join();
    }
    if (message.serverApi !== undefined) {
      ApiVersion.encode(message.serverApi, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApiCompatibilityError {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApiCompatibilityError();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.clientApi = ApiVersion.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.serverApi = ApiVersion.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<ApiCompatibilityError>): ApiCompatibilityError {
    return ApiCompatibilityError.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ApiCompatibilityError>): ApiCompatibilityError {
    const message = createBaseApiCompatibilityError();
    message.clientApi = (object.clientApi !== undefined && object.clientApi !== null)
      ? ApiVersion.fromPartial(object.clientApi)
      : undefined;
    message.serverApi = (object.serverApi !== undefined && object.serverApi !== null)
      ? ApiVersion.fromPartial(object.serverApi)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}